Convert oe_bootstrap_theme:create-subtheme to Drush.
